Chapter 9:  event-handling and destructuring
There are some examples of destructuring assignment notation which
Morten Rand-Hendricksen didn't explicitly talk about earlier
in the course; at least, I didn't remember him doing so.

So, there is a bit about that in this commit, and about the use
of the implicit "this" being available in a callback function!